+++ Frameworks/MAC/mac-src/src/MACLib/Old/APEDecompressOld.cpp
@@ -62,7 +62,7 @@ int CAPEDecompressOld::InitializeDecompressor()
     return Seek(0);
 }
 
-int CAPEDecompressOld::GetData(char * pBuffer, int nBlocks, int * pBlocksRetrieved)
+int CAPEDecompressOld::GetData(char * pBuffer, intptr_t nBlocks, int * pBlocksRetrieved)
 {
     if (pBlocksRetrieved) *pBlocksRetrieved = 0;
 
@@ -250,7 +250,7 @@ int CAPEDecompressOld::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nPa
             }
             else
             {
-                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (int) &wfeFormat, 0);
+                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (intptr_t) &wfeFormat, 0);
                 WAVE_HEADER WAVHeader; FillWaveHeader(&WAVHeader, 
                     (m_nFinishBlock - m_nStartBlock) * GetInfo(APE_INFO_BLOCK_ALIGN), 
                     &wfeFormat,    0);
